+++ ext/scripts/popup.js
@@ -3,6 +3,21 @@
 // Functions exclusive to the RetroTxt popup.html.
 // These run in isolation to the other context-scripts.
 
+/**
+ * Create a link to the extension Details tab.
+ * This is a clone of the same function found in helpers. */
+function LinkDetails() {
+  const extensionId = chrome.runtime.id,
+    ua = navigator.userAgent
+  if (extensionId.length === 0) return ``
+  if (ua.includes(`Firefox/`)) return ``
+  const url = `://extensions?id=${extensionId}`
+  if (ua.includes(`Edg/`)) return `edge${url}`
+  if (ua.includes(`OPR/`)) return `opera${url}`
+  // brave, vivaldi do not modify the user agent and cannot be detected
+  return `chrome${url}`
+}
+
 chrome.tabs.query({ currentWindow: true, active: true }, () => {
   // list website domains
   chrome.storage.local.get(`settingsWebsiteDomains`, (store) => {
@@ -20,8 +35,16 @@ chrome.tabs.query({ currentWindow: true, active: true }, () => {
   // is file scheme access allowed?
   chrome.extension.isAllowedFileSchemeAccess((allowed) => {
     if (!allowed) {
-      const div = document.getElementById(`localfiles`)
-      div.textContent = `When allow access to file URLs is toggled, ${div.textContent}`
+      const div1 = document.getElementById(`localfiles`)
+      div1.classList.add(`is-hidden`)
+      const div2 = document.getElementById(`unlockLocalfiles`)
+      div2.classList.remove(`is-hidden`)
+      const a = document.getElementById(`linkUnlock`)
+      a.addEventListener(`click`, () => {
+        chrome.tabs.create({
+          url: LinkDetails(),
+        })
+      })
     }
   })